feat: add new page component for listings
This commit adds a new page component called PageComponent.razor that will be used to display listings. It includes the necessary HTML and CSS code to render the listings in a card format. The component also makes an HTTP request to retrieve the listings data from the server and displays it dynamically.